refactor(ci): split latency evidence helpers
