+++ .buildchain/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.4",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.4",
-    "sourceSha": "6c44ce3f9340cb78d281fc4d4148f2e56edb930f"
+    "sourceSha": "6c6d50671fbf364e93326286a22fb9d664f32482"
   },
   "claims": [
     {
